Move many common utilities from regression_tests/common.py to libensemble/utils.py
