+++ libensemble/tests/regression_tests/script_test_balsam_hworld.py
@@ -11,7 +11,7 @@ from libensemble.message_numbers import WORKER_DONE, WORKER_KILL_ON_ERR, WORKER_
 from libensemble.libE import libE
 from libensemble.sim_funcs.job_control_hworld import job_control_hworld as sim_f
 from libensemble.gen_funcs.sampling import uniform_random_sample as gen_f
-from libensemble.tests.regression_tests.common import per_worker_stream
+from libensemble.util import per_worker_stream
 
 mpi4py.rc.recv_mprobe = False  # Disable matching probes
