+++ libensemble/tests/regression_tests/test_6-hump_camel_aposmm_LD_MMA.py
@@ -22,7 +22,7 @@ from libensemble.libE import libE, libE_tcp_worker
 from libensemble.sim_funcs.six_hump_camel import six_hump_camel as sim_f
 from libensemble.gen_funcs.aposmm import aposmm_logic as gen_f
 from libensemble.alloc_funcs.fast_alloc_to_aposmm import give_sim_work_first as alloc_f
-from libensemble.tests.regression_tests.common import parse_args, save_libE_output, per_worker_stream
+from libensemble.util import parse_args, save_libE_output, per_worker_stream
 from libensemble.tests.regression_tests.support import persis_info_1 as persis_info, aposmm_gen_out as gen_out, six_hump_camel_minima as minima
 
 nworkers, is_master, libE_specs, _ = parse_args()
