+++ libensemble/tests/regression_tests/test_6-hump_camel_elapsed_time_abort.py
@@ -21,7 +21,7 @@ from libensemble.libE import libE
 from libensemble.sim_funcs.six_hump_camel import six_hump_camel as sim_f
 from libensemble.gen_funcs.sampling import uniform_random_sample as gen_f
 from libensemble.alloc_funcs.give_sim_work_first import give_sim_work_first
-from libensemble.tests.regression_tests.common import parse_args, save_libE_output, per_worker_stream, eprint
+from libensemble.util import parse_args, save_libE_output, per_worker_stream, eprint
 
 nworkers, is_master, libE_specs, _ = parse_args()
