+++ libensemble/tests/regression_tests/test_calc_exception.py
@@ -18,7 +18,7 @@ import numpy as np
 from libensemble.libE import libE
 from libensemble.libE_manager import ManagerException
 from libensemble.gen_funcs.sampling import uniform_random_sample as gen_f
-from libensemble.tests.regression_tests.common import parse_args, per_worker_stream
+from libensemble.util import parse_args, per_worker_stream
 
 nworkers, is_master, libE_specs, _ = parse_args()
