+++ libensemble/tests/regression_tests/test_chwirut_uniform_sampling_one_residual_at_a_time.py
@@ -27,7 +27,7 @@ from libensemble.sim_funcs.chwirut1 import chwirut_eval as sim_f
 from libensemble.gen_funcs.sampling import uniform_random_sample_obj_components as gen_f
 from libensemble.alloc_funcs.fast_alloc_and_pausing import give_sim_work_first as alloc_f
 from libensemble.tests.regression_tests.support import persis_info_3 as persis_info
-from libensemble.tests.regression_tests.common import parse_args, save_libE_output, per_worker_stream
+from libensemble.util import parse_args, save_libE_output, per_worker_stream
 
 nworkers, is_master, libE_specs, _ = parse_args()
 if libE_specs['comms'] == 'tcp':
