+++ libensemble/tests/regression_tests/test_comms.py
@@ -20,7 +20,7 @@ import numpy as np
 from libensemble.libE import libE
 from libensemble.sim_funcs.comms_testing import float_x1000 as sim_f
 from libensemble.gen_funcs.sampling import uniform_random_sample as gen_f
-from libensemble.tests.regression_tests.common import parse_args, save_libE_output, per_worker_stream
+from libensemble.util import parse_args, save_libE_output, per_worker_stream
 from libensemble.mpi_controller import MPIJobController  # Only used to get workerID in float_x1000
 jobctrl = MPIJobController(auto_resources=False)
