+++ libensemble/tests/regression_tests/test_jobcontroller_hworld.py
@@ -21,7 +21,7 @@ from libensemble.message_numbers import WORKER_DONE, WORKER_KILL_ON_ERR, WORKER_
 from libensemble.libE import libE
 from libensemble.sim_funcs.job_control_hworld import job_control_hworld as sim_f
 from libensemble.gen_funcs.sampling import uniform_random_sample as gen_f
-from libensemble.tests.regression_tests.common import build_simfunc, parse_args, per_worker_stream
+from libensemble.util import build_simfunc, parse_args, per_worker_stream
 
 # Do not change these lines - they are parsed by run-tests.sh
 # TESTSUITE_COMMS: mpi local tcp
