+++ libensemble/tests/regression_tests/test_mpi_comms.py
@@ -10,7 +10,7 @@
 import sys
 from mpi4py import MPI
 from libensemble.comms.mpi import MPIComm, Timeout
-from libensemble.tests.regression_tests.common import parse_args
+from libensemble.util import parse_args
 
 # Do not change these lines - they are parsed by run-tests.sh
 # TESTSUITE_COMMS: mpi
