+++ libensemble/tests/regression_tests/test_worker_exceptions.py
@@ -19,7 +19,7 @@ from libensemble.libE import libE
 from libensemble.tests.regression_tests.support import nan_func as sim_f
 from libensemble.gen_funcs.sampling import uniform_random_sample as gen_f
 from libensemble.libE_manager import ManagerException
-from libensemble.tests.regression_tests.common import parse_args, per_worker_stream
+from libensemble.util import parse_args, per_worker_stream
 
 nworkers, is_master, libE_specs, _ = parse_args()
 n = 2
